+++ Flow/Changes.md
@@ -16,6 +16,10 @@ collection types when used.
 ### New Methods
 - CollectionExtensions
     - Added .slice(Range) to SeqOps
+    
+### Other Changes
+- ModelDeclaration.validate now makes sure the required values are non-empty. This applies for 
+String, Vector and Model type value requirements.
 
 ## v1.7